Initial commit to the repo
Yes, this is a pretty messy set of code. However, I'm okay with that, as
I was experimenting and learning as I was writing the code. So, in a
way, it was going to be messy. From here, however, the code is going to
be progressively refactored and cleaned up until it's something that I
would be happy putting my name to.